requirements.txt
tvdatafeed
psycopg2-binary
pandas
numpy
python-dotenv
sqlalchemy
ta-lib
